+++ instruction/srai.go
@@ -21,7 +21,7 @@ var SRAI = Instruction{
 		op1 := _inst.p.ReadReg(_inst.rs1)                            // first operator of alu
 		op2 := _inst.shamt                                           // second operator of alu
 		result := bv.Bv(32)                                          // will hold the result of alu
-		result.From(uint64(int64(op1.ToUint32()) >> op2.ToUint32())) // perform computation
+		result.From(uint32(int32(op1.ToUint32()) >> op2.ToUint32())) // perform computation
 		_inst.p.WriteReg(_inst.rd, result)                           // write result back to rd
 	},
 }
